refactor: fix the typo in the generate-title module filename
web/lib/genarate-title.ts was misspelled. Rename it and update the 43
import sites. The exported symbol generateMetaTitle was already spelled
correctly, so only the filename and the @/lib specifier change; the file
content is byte-identical.
