+++ web/app/(portal)/teams/[teamId]/(team)/app/page.tsx
@@ -1,4 +1,4 @@
-import { generateMetaTitle } from "@/lib/genarate-title";
+import { generateMetaTitle } from "@/lib/generate-title";
 import { urls } from "@/lib/urls";
 import { Metadata } from "next";
 import { redirect } from "next/navigation";
