+++ web/app/(portal)/teams/[teamId]/(team)/page.tsx
@@ -1,4 +1,4 @@
-import { generateMetaTitle } from "@/lib/genarate-title";
+import { generateMetaTitle } from "@/lib/generate-title";
 import { TeamIdPage } from "@/scenes/PortalV3/Teams/TeamId/Team/page";
 import { Metadata } from "next";
