+++ web/app/(portal)/teams/[teamId]/apps/[appId]/actions/[actionId]/proof-debugging/page.tsx
@@ -1,4 +1,4 @@
-import { generateMetaTitle } from "@/lib/genarate-title";
+import { generateMetaTitle } from "@/lib/generate-title";
 import { ActionIdProofDebugingPage } from "@/scenes/PortalV3/Teams/TeamId/Apps/AppId/Actions/ActionId/ProofDebuging/page";
 import { Metadata } from "next";
