+++ web/app/(portal)/teams/[teamId]/apps/[appId]/mini-app/notifications/page.tsx
@@ -1,4 +1,4 @@
-import { generateMetaTitle } from "@/lib/genarate-title";
+import { generateMetaTitle } from "@/lib/generate-title";
 import { NotificationsPage } from "@/scenes/PortalV3/Teams/TeamId/Apps/AppId/MiniApp/Notifications/page";
 import { Metadata } from "next";
