+++ web/app/(portal)/teams/[teamId]/apps/[appId]/mini-app/permissions/page.tsx
@@ -1,4 +1,4 @@
-import { generateMetaTitle } from "@/lib/genarate-title";
+import { generateMetaTitle } from "@/lib/generate-title";
 import { urls } from "@/lib/urls";
 import { Metadata } from "next";
 import { redirect } from "next/navigation";
