+++ web/app/admin/apps/[id]/page.tsx
@@ -1,5 +1,5 @@
 import { requireAdminUser } from "@/lib/admin-auth";
-import { generateMetaTitle } from "@/lib/genarate-title";
+import { generateMetaTitle } from "@/lib/generate-title";
 import { AdminAppPage } from "@/scenes/Admin/apps/id/page";
 import type { Metadata } from "next";
