+++ web/app/admin/rps/[id]/page.tsx
@@ -1,5 +1,5 @@
 import { requireAdminUser } from "@/lib/admin-auth";
-import { generateMetaTitle } from "@/lib/genarate-title";
+import { generateMetaTitle } from "@/lib/generate-title";
 import { AdminRpPage } from "@/scenes/Admin/rps/id/page";
 import type { Metadata } from "next";
