+++ web/app/admin/rps/page.tsx
@@ -1,5 +1,5 @@
 import { requireAdminUser } from "@/lib/admin-auth";
-import { generateMetaTitle } from "@/lib/genarate-title";
+import { generateMetaTitle } from "@/lib/generate-title";
 import { AdminRpsPage } from "@/scenes/Admin/rps/page";
 import { Metadata } from "next";
